pyproject.toml
[project]
name = "AutomobileFraudDetection"
version = "0.1.0"
description = "Automobile fraud detection with machine learning on graphs"
authors = [
    {name = "Matej Kalc", email = "matejkalc99@protonmail.com"},
]
dependencies = [
    "pandas>=2.2.1",
    "numpy>=1.26.4",
    "matplotlib<=3.9.0",
    "scikit-learn>=1.5.1",
    "xgboost>=2.0.3",
    "torch>=2.2.2",
    "networkx>=3.3",
    "torch-geometric>=2.5.2",
    "chardet>=5.2.0",
    "pyarrow>=15.0.2",
    "python-dotenv>=1.0.1",
    "scikit-optimize>=0.10.1",
    "imblearn>=0.0",
    "pyod>=1.1.3",
    "mlflow>=2.12.2",
    "rstcheck>=6.2.1",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "mypy>=1.9.0",
    "ruff>=0.3.5",
    "pre-commit>=3.7.0",
    "pre-commit-hooks>=4.6.0",
    "pandas-stubs>=2.2.1.240316",
    "types-PyYAML>=6.0.12.20240311",
    "types-requests>=2.31.0.20240406",
    "types-tqdm>=4.66.0.20240417",
]
[tool.ruff]

line-length = 140

[tool.ruff.lint]

extend-select = ["ALL"]

ignore = ["S301", "D104", "G004", "TRY003", "EM101", "EM102"]

[tool.ruff.lint.per-file-ignores]
"**/{tests}/*" = ["S101", "INP001"]
"**/{docs}/*" = ["INP001"]

[tool.ruff.lint.pydocstyle]
convention = "google"


